create debug.c and move debug code there
Hide the debug implementation from the rest of the kernel.
