+++ src/kernel/faulthandler.c
@@ -4,19 +4,24 @@
  * SPDX-License-Identifier: GPL-2.0-only
  */
 
+#include <config.h>
 #include <api/failures.h>
+#include <api/debug.h>
 #include <kernel/cspace.h>
 #include <kernel/faulthandler.h>
 #include <kernel/thread.h>
-#include <machine/io.h>
 #include <arch/machine.h>
 
 #ifdef CONFIG_KERNEL_MCS
+
 void handleFault(tcb_t *tptr)
 {
     bool_t hasFaultHandler = sendFaultIPC(tptr, TCB_PTR_CTE_PTR(tptr, tcbFaultHandler)->cap,
                                           tptr->tcbSchedContext != NULL);
     if (!hasFaultHandler) {
+#ifdef CONFIG_PRINTING
+        debug_print_fault_handler(tptr, current_fault);
+#endif /* CONFIG_PRINTING */
         handleNoFaultHandler(tptr);
     }
 }
@@ -48,15 +53,24 @@ bool_t sendFaultIPC(tcb_t *tptr, cap_t handlerCap, bool_t can_donate)
         return false;
     }
 }
-#else
+
+void handleNoFaultHandler(tcb_t *tptr)
+{
+    /* Debug output is done where this is called. */
+    setThreadState(tptr, ThreadState_Inactive);
+}
+
+#else /* not CONFIG_KERNEL_MCS */
 
 void handleFault(tcb_t *tptr)
 {
-    exception_t status;
     seL4_Fault_t fault = current_fault;
-
-    status = sendFaultIPC(tptr);
+    exception_t status = sendFaultIPC(tptr);
     if (status != EXCEPTION_NONE) {
+#ifdef CONFIG_PRINTING
+        /* current_fault holds the fault from the failed sendFaultIPC() now. */
+        debug_print_fault_handler(tptr, fault, status, current_fault);
+#endif /* CONFIG_PRINTING */
         handleDoubleFault(tptr, fault);
     }
 }
@@ -99,73 +113,11 @@ exception_t sendFaultIPC(tcb_t *tptr)
         return EXCEPTION_FAULT;
     }
 }
-#endif
 
-#ifdef CONFIG_PRINTING
-static void print_fault(seL4_Fault_t f)
-{
-    switch (seL4_Fault_get_seL4_FaultType(f)) {
-    case seL4_Fault_NullFault:
-        printf("null fault");
-        break;
-    case seL4_Fault_CapFault:
-        printf("cap fault in %s phase at address %p",
-               seL4_Fault_CapFault_get_inReceivePhase(f) ? "receive" : "send",
-               (void *)seL4_Fault_CapFault_get_address(f));
-        break;
-    case seL4_Fault_VMFault:
-        printf("vm fault on %s at address %p with status %p",
-               seL4_Fault_VMFault_get_instructionFault(f) ? "code" : "data",
-               (void *)seL4_Fault_VMFault_get_address(f),
-               (void *)seL4_Fault_VMFault_get_FSR(f));
-        break;
-    case seL4_Fault_UnknownSyscall:
-        printf("unknown syscall %p",
-               (void *)seL4_Fault_UnknownSyscall_get_syscallNumber(f));
-        break;
-    case seL4_Fault_UserException:
-        printf("user exception %p code %p",
-               (void *)seL4_Fault_UserException_get_number(f),
-               (void *)seL4_Fault_UserException_get_code(f));
-        break;
-#ifdef CONFIG_KERNEL_MCS
-    case seL4_Fault_Timeout:
-        printf("Timeout fault for 0x%x\n", (unsigned int) seL4_Fault_Timeout_get_badge(f));
-        break;
-#endif
-    default:
-        printf("unknown fault");
-        break;
-    }
-}
-#endif
-
-#ifdef CONFIG_KERNEL_MCS
-void handleNoFaultHandler(tcb_t *tptr)
-#else
-/* The second fault, ex2, is stored in the global current_fault */
 void handleDoubleFault(tcb_t *tptr, seL4_Fault_t ex1)
-#endif
 {
-#ifdef CONFIG_PRINTING
-#ifdef CONFIG_KERNEL_MCS
-    printf("Found thread has no fault handler while trying to handle:\n");
-    print_fault(current_fault);
-#else
-    seL4_Fault_t ex2 = current_fault;
-    printf("Caught ");
-    print_fault(ex2);
-    printf("\nwhile trying to handle:\n");
-    print_fault(ex1);
-#endif
-#ifdef CONFIG_DEBUG_BUILD
-    printf("\nin thread %p \"%s\" ", tptr, TCB_PTR_DEBUG_PTR(tptr)->tcbName);
-#endif /* CONFIG_DEBUG_BUILD */
-
-    printf("at address %p\n", (void *)getRestartPC(tptr));
-    printf("With stack:\n");
-    Arch_userStackTrace(tptr);
-#endif
-
+    /* debug output is done where this is called. */
     setThreadState(tptr, ThreadState_Inactive);
 }
+
+#endif /* [not] CONFIG_KERNEL_MCS */
